move omdb APIs and instance_migrate to lockstep API
