+++ dev-tools/omdb/Cargo.toml
@@ -37,10 +37,12 @@ http.workspace = true
 humantime.workspace = true
 iddqd.workspace = true
 indent_write.workspace = true
+indicatif.workspace = true
 internal-dns-resolver.workspace = true
 internal-dns-types.workspace = true
+ipnetwork.workspace = true
 itertools.workspace = true
-nexus-client.workspace = true
+multimap.workspace = true
 nexus-config.workspace = true
 nexus-db-errors.workspace = true
 nexus-db-lookup.workspace = true
@@ -48,15 +50,20 @@ nexus-db-model.workspace = true
 nexus-db-queries.workspace = true
 nexus-db-schema.workspace = true
 nexus-inventory.workspace = true
+nexus-lockstep-client.workspace = true
 nexus-reconfigurator-preparation.workspace = true
 nexus-saga-recovery.workspace = true
 nexus-sled-agent-shared.workspace = true
 nexus-types.workspace = true
 omicron-common.workspace = true
 omicron-uuid-kinds.workspace = true
+omicron-workspace-hack.workspace = true
+owo-colors.workspace = true
 oxide-tokio-rt.workspace = true
 oximeter-client.workspace = true
 oximeter-db = { workspace = true, default-features = false, features = [ "oxql" ] }
+oxnet.workspace = true
+petgraph.workspace = true
 # See omicron-rpaths for more about the "pq-sys" dependency.
 pq-sys = "*"
 ratatui.workspace = true
@@ -79,13 +86,6 @@ unicode-width.workspace = true
 update-engine.workspace = true
 url.workspace = true
 uuid.workspace = true
-ipnetwork.workspace = true
-omicron-workspace-hack.workspace = true
-multimap.workspace = true
-indicatif.workspace = true
-petgraph.workspace = true
-oxnet.workspace = true
-owo-colors.workspace = true
 
 [dev-dependencies]
 camino-tempfile.workspace = true
