+++ dev-tools/omdb/src/bin/omdb/nexus.rs
@@ -31,19 +31,19 @@ use futures::TryStreamExt;
 use http::StatusCode;
 use internal_dns_types::names::ServiceName;
 use itertools::Itertools;
-use nexus_client::types::ActivationReason;
-use nexus_client::types::BackgroundTask;
-use nexus_client::types::BackgroundTasksActivateRequest;
-use nexus_client::types::CurrentStatus;
-use nexus_client::types::LastResult;
-use nexus_client::types::PhysicalDiskPath;
-use nexus_client::types::SagaState;
-use nexus_client::types::SledSelector;
-use nexus_client::types::UninitializedSledId;
 use nexus_db_lookup::LookupPath;
 use nexus_db_queries::context::OpContext;
 use nexus_db_queries::db::DataStore;
 use nexus_inventory::now_db_precision;
+use nexus_lockstep_client::types::ActivationReason;
+use nexus_lockstep_client::types::BackgroundTask;
+use nexus_lockstep_client::types::BackgroundTasksActivateRequest;
+use nexus_lockstep_client::types::CurrentStatus;
+use nexus_lockstep_client::types::LastResult;
+use nexus_lockstep_client::types::PhysicalDiskPath;
+use nexus_lockstep_client::types::SagaState;
+use nexus_lockstep_client::types::SledSelector;
+use nexus_lockstep_client::types::UninitializedSledId;
 use nexus_saga_recovery::LastPass;
 use nexus_types::deployment::Blueprint;
 use nexus_types::deployment::ClickhouseMode;
@@ -113,7 +113,7 @@ use uuid::Uuid;
 /// Arguments to the "omdb nexus" subcommand
 #[derive(Debug, Args)]
 pub struct NexusArgs {
-    /// URL of the Nexus internal API
+    /// URL of the Nexus internal lockstep API
     #[clap(
         long,
         env = "OMDB_NEXUS_URL",
@@ -268,7 +268,7 @@ impl BlueprintIdOrCurrentTarget {
 
     async fn resolve_to_id_via_nexus(
         &self,
-        client: &nexus_client::Client,
+        client: &nexus_lockstep_client::Client,
     ) -> anyhow::Result<BlueprintUuid> {
         match self {
             Self::CurrentTarget => {
@@ -284,7 +284,7 @@ impl BlueprintIdOrCurrentTarget {
 
     async fn resolve_to_blueprint(
         &self,
-        client: &nexus_client::Client,
+        client: &nexus_lockstep_client::Client,
     ) -> anyhow::Result<Blueprint> {
         let id = self.resolve_to_id_via_nexus(client).await?;
         let response = client
@@ -614,13 +614,14 @@ impl NexusArgs {
                     "note: Nexus URL not specified.  Will pick one from DNS."
                 );
                 let addr = omdb
-                    .dns_lookup_one(log.clone(), ServiceName::Nexus)
+                    .dns_lookup_one(log.clone(), ServiceName::NexusLockstep)
                     .await?;
                 format!("http://{}", addr)
             }
         };
         eprintln!("note: using Nexus URL {}", &nexus_url);
-        let client = nexus_client::Client::new(&nexus_url, log.clone());
+        let client =
+            nexus_lockstep_client::Client::new(&nexus_url, log.clone());
 
         match &self.command {
             NexusCommands::BackgroundTasks(BackgroundTasksArgs {
@@ -837,7 +838,7 @@ impl NexusArgs {
 
 /// Runs `omdb nexus background-tasks doc`
 async fn cmd_nexus_background_tasks_doc(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
 ) -> Result<(), anyhow::Error> {
     let response =
         client.bgtask_list().await.context("listing background tasks")?;
@@ -863,7 +864,7 @@ async fn cmd_nexus_background_tasks_doc(
 
 /// Runs `omdb nexus background-tasks list`
 async fn cmd_nexus_background_tasks_list(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
 ) -> Result<(), anyhow::Error> {
     let response =
         client.bgtask_list().await.context("listing background tasks")?;
@@ -881,7 +882,7 @@ async fn cmd_nexus_background_tasks_list(
 
 /// Runs `omdb nexus background-tasks show`
 async fn cmd_nexus_background_tasks_show(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &BackgroundTasksShowArgs,
 ) -> Result<(), anyhow::Error> {
     let response =
@@ -966,7 +967,7 @@ async fn cmd_nexus_background_tasks_show(
 
 /// Runs `omdb nexus background-tasks print-report`
 async fn cmd_nexus_background_tasks_print_report(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &BackgroundTasksPrintReportArgs,
     color: ColorChoice,
 ) -> Result<(), anyhow::Error> {
@@ -1009,7 +1010,7 @@ async fn cmd_nexus_background_tasks_print_report(
 
 /// Runs `omdb nexus background-tasks activate`
 async fn cmd_nexus_background_tasks_activate(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &BackgroundTasksActivateArgs,
     // This isn't quite "destructive" in the sense that of it being potentially
     // dangerous, but it does modify the system rather than being a read-only
@@ -3240,7 +3241,7 @@ fn reason_code(reason: ActivationReason) -> char {
 }
 
 async fn cmd_nexus_blueprints_list(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
 ) -> Result<(), anyhow::Error> {
     #[derive(Tabled)]
     #[tabled(rename_all = "SCREAMING_SNAKE_CASE")]
@@ -3312,7 +3313,7 @@ async fn cmd_nexus_blueprints_list(
 }
 
 async fn cmd_nexus_blueprints_show(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &BlueprintIdArgs,
 ) -> Result<(), anyhow::Error> {
     let blueprint = args.blueprint_id.resolve_to_blueprint(client).await?;
@@ -3321,7 +3322,7 @@ async fn cmd_nexus_blueprints_show(
 }
 
 async fn cmd_nexus_blueprints_diff(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &BlueprintDiffArgs,
 ) -> Result<(), anyhow::Error> {
     let blueprint = args.blueprint1_id.resolve_to_blueprint(client).await?;
@@ -3348,7 +3349,7 @@ async fn cmd_nexus_blueprints_diff(
 }
 
 async fn cmd_nexus_blueprints_delete(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &BlueprintIdArgs,
     _destruction_token: DestructiveOperationToken,
 ) -> Result<(), anyhow::Error> {
@@ -3363,7 +3364,7 @@ async fn cmd_nexus_blueprints_delete(
 }
 
 async fn cmd_nexus_blueprints_target_show(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
 ) -> Result<(), anyhow::Error> {
     let target = client
         .blueprint_target_view()
@@ -3376,7 +3377,7 @@ async fn cmd_nexus_blueprints_target_show(
 }
 
 async fn cmd_nexus_blueprints_target_set(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &BlueprintTargetSetArgs,
     _destruction_token: DestructiveOperationToken,
 ) -> Result<(), anyhow::Error> {
@@ -3432,10 +3433,12 @@ async fn cmd_nexus_blueprints_target_set(
     };
 
     client
-        .blueprint_target_set(&nexus_client::types::BlueprintTargetSet {
-            target_id: args.blueprint_id,
-            enabled,
-        })
+        .blueprint_target_set(
+            &nexus_lockstep_client::types::BlueprintTargetSet {
+                target_id: args.blueprint_id,
+                enabled,
+            },
+        )
         .await
         .with_context(|| {
             format!("setting target to blueprint {}", args.blueprint_id)
@@ -3445,7 +3448,7 @@ async fn cmd_nexus_blueprints_target_set(
 }
 
 async fn cmd_nexus_blueprints_target_set_enabled(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &BlueprintIdArgs,
     enabled: bool,
     _destruction_token: DestructiveOperationToken,
@@ -3455,7 +3458,7 @@ async fn cmd_nexus_blueprints_target_set_enabled(
     let description = if enabled { "enabled" } else { "disabled" };
     client
         .blueprint_target_set_enabled(
-            &nexus_client::types::BlueprintTargetSet {
+            &nexus_lockstep_client::types::BlueprintTargetSet {
                 target_id: blueprint_id,
                 enabled,
             },
@@ -3469,7 +3472,7 @@ async fn cmd_nexus_blueprints_target_set_enabled(
 }
 
 async fn cmd_nexus_blueprints_regenerate(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     _destruction_token: DestructiveOperationToken,
 ) -> Result<(), anyhow::Error> {
     let blueprint =
@@ -3479,7 +3482,7 @@ async fn cmd_nexus_blueprints_regenerate(
 }
 
 async fn cmd_nexus_blueprints_import(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     _destruction_token: DestructiveOperationToken,
     args: &BlueprintImportArgs,
 ) -> Result<(), anyhow::Error> {
@@ -3497,7 +3500,7 @@ async fn cmd_nexus_blueprints_import(
 }
 
 async fn cmd_nexus_clickhouse_policy_get(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
 ) -> Result<(), anyhow::Error> {
     let res = client.clickhouse_policy_get().await;
 
@@ -3541,7 +3544,7 @@ async fn cmd_nexus_clickhouse_policy_get(
 }
 
 async fn cmd_nexus_mgs_updates(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
 ) -> Result<(), anyhow::Error> {
     let response = client
         .mgs_updates()
@@ -3553,7 +3556,7 @@ async fn cmd_nexus_mgs_updates(
 }
 
 async fn cmd_nexus_clickhouse_policy_set(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &ClickhousePolicySetArgs,
     _destruction_token: DestructiveOperationToken,
 ) -> Result<(), anyhow::Error> {
@@ -3603,7 +3606,7 @@ async fn cmd_nexus_clickhouse_policy_set(
 }
 
 async fn cmd_nexus_oximeter_read_policy_get(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
 ) -> Result<(), anyhow::Error> {
     let res = client.oximeter_read_policy_get().await;
 
@@ -3637,7 +3640,7 @@ async fn cmd_nexus_oximeter_read_policy_get(
 }
 
 async fn cmd_nexus_oximeter_read_policy_set(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &OximeterReadPolicySetArgs,
     _destruction_token: DestructiveOperationToken,
 ) -> Result<(), anyhow::Error> {
@@ -3681,7 +3684,7 @@ async fn cmd_nexus_oximeter_read_policy_set(
 
 /// Runs `omdb nexus sagas list`
 async fn cmd_nexus_sagas_list(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
 ) -> Result<(), anyhow::Error> {
     // We don't want users to confuse this with a general way to list all sagas.
     // Such a command would read database state and it would go under "omdb db".
@@ -3726,7 +3729,7 @@ async fn cmd_nexus_sagas_list(
 
 /// Runs `omdb nexus sagas demo-create`
 async fn cmd_nexus_sagas_demo_create(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     _destruction_token: DestructiveOperationToken,
 ) -> Result<(), anyhow::Error> {
     let demo_saga =
@@ -3741,7 +3744,7 @@ async fn cmd_nexus_sagas_demo_create(
 
 /// Runs `omdb nexus sagas demo-complete`
 async fn cmd_nexus_sagas_demo_complete(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &DemoSagaIdArgs,
     _destruction_token: DestructiveOperationToken,
 ) -> Result<(), anyhow::Error> {
@@ -3766,7 +3769,7 @@ async fn cmd_nexus_sagas_demo_complete(
 
 /// Runs `omdb nexus sleds list-uninitialized`
 async fn cmd_nexus_sleds_list_uninitialized(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
 ) -> Result<(), anyhow::Error> {
     let response = client
         .sled_list_uninitialized()
@@ -3808,7 +3811,7 @@ async fn cmd_nexus_sleds_list_uninitialized(
 
 /// Runs `omdb nexus sleds add`
 async fn cmd_nexus_sled_add(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &SledAddArgs,
     _destruction_token: DestructiveOperationToken,
 ) -> Result<(), anyhow::Error> {
@@ -3827,7 +3830,7 @@ async fn cmd_nexus_sled_add(
 
 /// Runs `omdb nexus sleds expunge`
 async fn cmd_nexus_sled_expunge(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &SledExpungeArgs,
     omdb: &Omdb,
     log: &slog::Logger,
@@ -3849,7 +3852,7 @@ async fn cmd_nexus_sled_expunge(
 // `omdb nexus sleds expunge`, but borrowing a datastore
 async fn cmd_nexus_sled_expunge_with_datastore(
     datastore: &Arc<DataStore>,
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &SledExpungeArgs,
     log: &slog::Logger,
     _destruction_token: DestructiveOperationToken,
@@ -3938,7 +3941,7 @@ async fn cmd_nexus_sled_expunge_with_datastore(
 
 /// Runs `omdb nexus sleds expunge-disk`
 async fn cmd_nexus_sled_expunge_disk(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &DiskExpungeArgs,
     omdb: &Omdb,
     log: &slog::Logger,
@@ -3959,7 +3962,7 @@ async fn cmd_nexus_sled_expunge_disk(
 
 async fn cmd_nexus_sled_expunge_disk_with_datastore(
     datastore: &Arc<DataStore>,
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &DiskExpungeArgs,
     log: &slog::Logger,
     _destruction_token: DestructiveOperationToken,
@@ -4072,7 +4075,7 @@ async fn cmd_nexus_sled_expunge_disk_with_datastore(
 
 /// Runs `omdb nexus support-bundles list`
 async fn cmd_nexus_support_bundles_list(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
 ) -> Result<(), anyhow::Error> {
     let support_bundle_stream = client.support_bundle_list_stream(None, None);
 
@@ -4111,13 +4114,15 @@ async fn cmd_nexus_support_bundles_list(
 
 /// Runs `omdb nexus support-bundles create`
 async fn cmd_nexus_support_bundles_create(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     _destruction_token: DestructiveOperationToken,
 ) -> Result<(), anyhow::Error> {
     let support_bundle_id = client
-        .support_bundle_create(&nexus_client::types::SupportBundleCreate {
-            user_comment: None,
-        })
+        .support_bundle_create(
+            &nexus_lockstep_client::types::SupportBundleCreate {
+                user_comment: None,
+            },
+        )
         .await
         .context("creating support bundle")?
         .into_inner()
@@ -4128,7 +4133,7 @@ async fn cmd_nexus_support_bundles_create(
 
 /// Runs `omdb nexus support-bundles delete`
 async fn cmd_nexus_support_bundles_delete(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &SupportBundleDeleteArgs,
     _destruction_token: DestructiveOperationToken,
 ) -> Result<(), anyhow::Error> {
@@ -4158,7 +4163,7 @@ async fn write_stream_to_sink(
 //
 // "range" is in bytes, and is inclusive on both sides.
 async fn support_bundle_download_range(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     id: SupportBundleUuid,
     range: (u64, u64),
 ) -> anyhow::Result<impl futures::Stream<Item = anyhow::Result<bytes::Bytes>>> {
@@ -4177,7 +4182,7 @@ async fn support_bundle_download_range(
 // Starts the download at "start" bytes (inclusive) and continues up to "end"
 // bytes (exclusive).
 fn support_bundle_download_ranges(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     id: SupportBundleUuid,
     start: u64,
     end: u64,
@@ -4206,7 +4211,7 @@ fn support_bundle_download_ranges(
 
 /// Runs `omdb nexus support-bundles download`
 async fn cmd_nexus_support_bundles_download(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &SupportBundleDownloadArgs,
 ) -> Result<(), anyhow::Error> {
     let total_length = client
@@ -4244,7 +4249,7 @@ async fn cmd_nexus_support_bundles_download(
 
 /// Runs `omdb nexus support-bundles get-index`
 async fn cmd_nexus_support_bundles_get_index(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &SupportBundleIndexArgs,
 ) -> Result<(), anyhow::Error> {
     let stream = client
@@ -4264,7 +4269,7 @@ async fn cmd_nexus_support_bundles_get_index(
 
 /// Runs `omdb nexus support-bundles get-file`
 async fn cmd_nexus_support_bundles_get_file(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &SupportBundleFileArgs,
 ) -> Result<(), anyhow::Error> {
     let stream = client
@@ -4296,7 +4301,7 @@ async fn cmd_nexus_support_bundles_get_file(
 
 /// Runs `omdb nexus support-bundles inspect`
 async fn cmd_nexus_support_bundles_inspect(
-    client: &nexus_client::Client,
+    client: &nexus_lockstep_client::Client,
     args: &SupportBundleInspectArgs,
 ) -> Result<(), anyhow::Error> {
     let accessor: Box<dyn SupportBundleAccessor> = match (args.id, &args.path) {
