+++ dev-tools/omdb/tests/test_all_output.rs
@@ -136,8 +136,8 @@ async fn test_omdb_success_cases(cptestctx: &ControlPlaneTestContext) {
     let cmd_path = path_to_executable(CMD_OMDB);
 
     let postgres_url = cptestctx.database.listen_url();
-    let nexus_internal_url =
-        format!("http://{}/", cptestctx.internal_client.bind_address);
+    let nexus_lockstep_url =
+        format!("http://{}/", cptestctx.lockstep_client.bind_address);
     let mgs_url = cptestctx
         .gateway
         .get(&SwitchLocation::Switch0)
@@ -156,7 +156,7 @@ async fn test_omdb_success_cases(cptestctx: &ControlPlaneTestContext) {
     // Get the CockroachDB metadata from the blueprint so we can redact it
     let initial_blueprint: Blueprint = dropshot::test_util::read_json(
         &mut cptestctx
-            .internal_client
+            .lockstep_client
             .make_request_no_body(
                 Method::GET,
                 &format!("/deployment/blueprints/all/{initial_blueprint_id}"),
@@ -306,7 +306,7 @@ async fn test_omdb_success_cases(cptestctx: &ControlPlaneTestContext) {
     for args in invocations {
         println!("running commands with args: {:?}", args);
         let p = postgres_url.to_string();
-        let u = nexus_internal_url.clone();
+        let u = nexus_lockstep_url.clone();
         let g = mgs_url.clone();
         let ox = ox_url.clone();
         let ch = ch_url.clone();
@@ -396,8 +396,8 @@ async fn test_omdb_env_settings(cptestctx: &ControlPlaneTestContext) {
 
     let cmd_path = path_to_executable(CMD_OMDB);
     let postgres_url = cptestctx.database.listen_url().to_string();
-    let nexus_internal_url =
-        format!("http://{}", cptestctx.internal_client.bind_address);
+    let nexus_lockstep_url =
+        format!("http://{}", cptestctx.lockstep_client.bind_address);
     let ox_url = format!("http://{}/", cptestctx.oximeter.server_address());
     let ox_test_producer = cptestctx.producer.address().ip();
     let ch_url = format!("http://{}/", cptestctx.clickhouse.http_address());
@@ -425,7 +425,7 @@ async fn test_omdb_env_settings(cptestctx: &ControlPlaneTestContext) {
     let args = &[
         "nexus",
         "--nexus-internal-url",
-        &nexus_internal_url.clone(),
+        &nexus_lockstep_url.clone(),
         "background-tasks",
         "doc",
     ];
@@ -434,7 +434,7 @@ async fn test_omdb_env_settings(cptestctx: &ControlPlaneTestContext) {
     // Case 2: specified in multiple places (command-line argument wins)
     let args =
         &["nexus", "--nexus-internal-url", "junk", "background-tasks", "doc"];
-    let n = nexus_internal_url.clone();
+    let n = nexus_lockstep_url.clone();
     do_run(
         &mut output,
         move |exec| exec.env("OMDB_NEXUS_URL", &n),
