+++ live-tests/Cargo.toml
@@ -20,11 +20,11 @@ futures.workspace = true
 internal-dns-resolver.workspace = true
 internal-dns-types.workspace = true
 live-tests-macros.workspace = true
-nexus-client.workspace = true
 nexus-config.workspace = true
 nexus-db-model.workspace = true
 nexus-db-queries.workspace = true
 nexus-inventory.workspace = true
+nexus-lockstep-client.workspace = true
 nexus-reconfigurator-planning.workspace = true
 nexus-reconfigurator-preparation.workspace = true
 nexus-sled-agent-shared.workspace = true
