+++ live-tests/tests/test_nexus_handoff.rs
@@ -10,8 +10,8 @@ use anyhow::Context;
 use common::LiveTestContext;
 use common::reconfigurator::blueprint_edit_current_target;
 use live_tests_macros::live_test;
-use nexus_client::types::QuiesceState;
 use nexus_db_model::DbMetadataNexusState;
+use nexus_lockstep_client::types::QuiesceState;
 use nexus_reconfigurator_planning::blueprint_builder::BlueprintBuilder;
 use nexus_reconfigurator_preparation::PlanningInputFromDb;
 use nexus_types::deployment::BlueprintZoneDisposition;
