+++ nexus/Cargo.toml
@@ -151,14 +151,17 @@ criterion.workspace = true
 diesel.workspace = true
 dns-server.workspace = true
 expectorate.workspace = true
-hyper-rustls.workspace = true
 gateway-messages.workspace = true
 gateway-test-utils.workspace = true
+hickory-resolver.workspace = true
+httpmock.workspace = true
+httptest.workspace = true
 hubtools.workspace = true
+hyper-rustls.workspace = true
 nexus-db-queries = { workspace = true, features = ["testing"] }
-nexus-client.workspace = true
-nexus-test-utils-macros.workspace = true
+nexus-lockstep-client.workspace = true
 nexus-test-utils.workspace = true
+nexus-test-utils-macros.workspace = true
 omicron-sled-agent.workspace = true
 omicron-test-utils.workspace = true
 openapi-lint.workspace = true
@@ -170,18 +173,15 @@ petgraph.workspace = true
 pretty_assertions.workspace = true
 rcgen.workspace = true
 regex.workspace = true
+rustls.workspace = true
 similar-asserts.workspace = true
 sp-sim.workspace = true
-rustls.workspace = true
+strum.workspace = true
 subprocess.workspace = true
 term.workspace = true
-hickory-resolver.workspace = true
 tufaceous.workspace = true
-tufaceous-lib.workspace = true
-httptest.workspace = true
-httpmock.workspace = true
-strum.workspace = true
 tufaceous-artifact.workspace = true
+tufaceous-lib.workspace = true
 
 [[bench]]
 name = "setup_benchmark"
