+++ nexus/lockstep-api/Cargo.toml
@@ -9,5 +9,11 @@ workspace = true
 
 [dependencies]
 dropshot.workspace = true
+http.workspace = true
 nexus-types.workspace = true
+omicron-common.workspace = true
+omicron-uuid-kinds.workspace = true
 omicron-workspace-hack.workspace = true
+schemars.workspace = true
+serde.workspace = true
+uuid.workspace = true
