+++ nexus/src/app/quiesce.rs
@@ -517,9 +517,9 @@ mod test {
     use diesel::ExpressionMethods;
     use diesel::QueryDsl;
     use http::StatusCode;
-    use nexus_client::types::QuiesceState;
-    use nexus_client::types::QuiesceStatus;
     use nexus_db_model::DbMetadataNexusState;
+    use nexus_lockstep_client::types::QuiesceState;
+    use nexus_lockstep_client::types::QuiesceStatus;
     use nexus_test_interface::NexusServer;
     use nexus_test_utils::db::TestDatabase;
     use nexus_test_utils_macros::nexus_test;
@@ -540,11 +540,12 @@ mod test {
 
     type ControlPlaneTestContext =
         nexus_test_utils::ControlPlaneTestContext<crate::Server>;
-    type NexusClientError = nexus_client::Error<nexus_client::types::Error>;
+    type NexusClientError =
+        nexus_lockstep_client::Error<nexus_lockstep_client::types::Error>;
 
     async fn wait_quiesce(
         log: &Logger,
-        client: &nexus_client::Client,
+        client: &nexus_lockstep_client::Client,
         timeout: Duration,
     ) -> QuiesceStatus {
         wait_for_condition(
@@ -558,9 +559,7 @@ mod test {
                 if matches!(rv.state, QuiesceState::Quiesced { .. }) {
                     Ok(rv)
                 } else {
-                    Err(CondCheckError::<
-                        nexus_client::Error<nexus_client::types::Error>,
-                    >::NotYet)
+                    Err(CondCheckError::<NexusClientError>::NotYet)
                 }
             },
             &Duration::from_millis(50),
@@ -628,12 +627,14 @@ mod test {
     #[nexus_test(server = crate::Server)]
     async fn test_quiesce_easy(cptestctx: &ControlPlaneTestContext) {
         let log = &cptestctx.logctx.log;
-        let nexus_internal_url = format!(
+        let nexus_lockstep_url = format!(
             "http://{}",
-            cptestctx.server.get_http_server_internal_address().await
+            cptestctx.server.get_http_server_lockstep_address().await
         );
-        let nexus_client =
-            nexus_client::Client::new(&nexus_internal_url, log.clone());
+        let nexus_client = nexus_lockstep_client::Client::new(
+            &nexus_lockstep_url,
+            log.clone(),
+        );
 
         // We need to enable blueprint execution in order to complete a saga
         // assignment pass, which is required for quiescing to work.
@@ -657,12 +658,14 @@ mod test {
     #[nexus_test(server = crate::Server)]
     async fn test_quiesce_full(cptestctx: &ControlPlaneTestContext) {
         let log = &cptestctx.logctx.log;
-        let nexus_internal_url = format!(
+        let nexus_lockstep_url = format!(
             "http://{}",
-            cptestctx.server.get_http_server_internal_address().await
+            cptestctx.server.get_http_server_lockstep_address().await
         );
-        let nexus_client =
-            nexus_client::Client::new(&nexus_internal_url, log.clone());
+        let nexus_client = nexus_lockstep_client::Client::new(
+            &nexus_lockstep_url,
+            log.clone(),
+        );
 
         // We need to enable blueprint execution in order to complete a saga
         // assignment pass, which is required for quiescing to work.
