+++ nexus/tests/integration_tests/disks.rs
@@ -2348,8 +2348,8 @@ async fn test_disk_expunge(cptestctx: &ControlPlaneTestContext) {
     assert_eq!(allocated_regions.len(), REGION_REDUNDANCY_THRESHOLD);
 
     // Expunge the sled
-    let int_client = &cptestctx.internal_client;
-    int_client
+    cptestctx
+        .lockstep_client
         .make_request(
             Method::POST,
             "/sleds/expunge",
